+++ packages/react-native-meeting-starter-kit/src/screens/ActiveCallScreen.tsx
@@ -1,5 +1,9 @@
 import React, {useCallback} from 'react';
-import {ActiveCall, useCall} from '@stream-io/video-react-native-sdk';
+import {
+  ActiveCall,
+  StreamVideoRN,
+  useCall,
+} from '@stream-io/video-react-native-sdk';
 import {ActivityIndicator, StyleSheet, View} from 'react-native';
 import {NavigationStackParamsList} from '../types';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
@@ -19,14 +23,16 @@ export function ActiveCallScreen({navigation}: ActiveCallScreenProps) {
     navigation.navigate('CallParticipantsInfoScreen');
   }, [navigation]);
 
+  StreamVideoRN.setConfig({
+    onOpenCallParticipantsInfoView: onOpenCallParticipantsInfoViewHandler,
+  });
+
   if (!activeCall) {
     return <ActivityIndicator size={'large'} style={StyleSheet.absoluteFill} />;
   }
   return (
     <View style={[styles.wrapper, {paddingTop: insets.top}]}>
-      <ActiveCall
-        onOpenCallParticipantsInfoView={onOpenCallParticipantsInfoViewHandler}
-      />
+      <ActiveCall />
     </View>
   );
 }