+++ packages/styling/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.23](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.22...@stream-io/video-styling-0.0.1-alpha.23) (2023-05-17)
+
+
+
+## [0.0.1-alpha.22](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.21...@stream-io/video-styling-0.0.1-alpha.22) (2023-05-17)
+
+
+
+## [0.0.1-alpha.21](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.20...@stream-io/video-styling-0.0.1-alpha.21) (2023-05-16)
+
+
+
+## [0.0.1-alpha.20](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.19...@stream-io/video-styling-0.0.1-alpha.20) (2023-05-16)
+
+
+
+## [0.0.1-alpha.19](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.18...@stream-io/video-styling-0.0.1-alpha.19) (2023-05-16)
+
+
+
 ## [0.0.1-alpha.18](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.17...@stream-io/video-styling-0.0.1-alpha.18) (2023-05-16)